# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopmee API Documentation",
        "description": "API endpoints for Shopmee e-commerce application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Shopmee API Server"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng ký tài khoản người dùng mới (chưa kích hoạt)",
                "description": "Đăng ký thông tin tài khoản người dùng ban đầu. Sau khi thành công, hệ thống sẽ gửi mã OTP xác thực qua email.",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nguyen Van A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nva@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng ký thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng ký thông tin thành công! Vui lòng kiểm tra hộp thư Email để lấy mã OTP xác thực."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Lỗi validation (Ví dụ: Trùng Email, sai định dạng điện thoại...)"
                    },
                    "500": {
                        "description": "Lỗi hệ thống hoặc lỗi gửi mail"
                    }
                }
            }
        },
        "/register/verify": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Kích hoạt tài khoản bằng mã OTP",
                "description": "Nhập mã OTP được gửi qua email để hoàn tất việc đăng ký và nhận token đăng nhập.",
                "operationId": "verifyRegisterOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp_code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nva@example.com"
                                    },
                                    "otp_code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kích hoạt thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tài khoản của bạn đã được kích hoạt thành công!"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abcdefghijklmnop..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Mã OTP không chính xác hoặc hết hạn"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng nhập người dùng",
                "description": "Đăng nhập bằng Email và Password, trả về Access Token.",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nva@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dang nhap thanh cong!"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "1|abcdef..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Nguyen Van A"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "nva@example.com"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "buyer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Thông tin đăng nhập không chính xác"
                    }
                }
            }
        },
        "/password/forgot": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Gửi yêu cầu khôi phục mật khẩu (Quên mật khẩu)",
                "description": "Gửi mã OTP khôi phục mật khẩu tới địa chỉ Email đã đăng ký.",
                "operationId": "forgotPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nva@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gửi thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Hệ thống đã gửi mã OTP khôi phục mật khẩu vào Email của bạn."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/password/verify": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Kiểm tra mã OTP quên mật khẩu",
                "description": "Xác minh mã OTP khôi phục mật khẩu được gửi tới email. Trả về reset_token để đổi mật khẩu.",
                "operationId": "verifyPasswordOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp_code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nva@example.com"
                                    },
                                    "otp_code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Xác minh thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mã OTP chính xác. Vui lòng chuyển hướng sang trang đổi mật khẩu mới."
                                        },
                                        "reset_token": {
                                            "type": "string",
                                            "example": "reset_token_hash_value"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Mã OTP không đúng"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}